--- a/.gitignore
+++ b/.gitignore
@@ -1,4 +1,5 @@
 /_build
+/bench/snapshots
 /cover
 /doc
 /docs