--- a/lib/helpers/country.ex
+++ b/lib/helpers/country.ex
@@ -2,7 +2,8 @@ defmodule Helper.Country do
   @moduledoc false
   defmacro __using__(opts \\ []) do
     match_type = Keyword.get(opts, :match)
-    build_matcher(match_type)
+    number_prefix = Keyword.get(opts, :number_prefix, "")
+    build_matcher(match_type, number_prefix)
   end
 
   defmacro field(name, value) do
@@ -11,14 +12,15 @@ defmodule Helper.Country do
     end
   end
 
-  defp regex_matcher do
+  defp regex_matcher(number_prefix) do
     quote do
       import Helper.Country
       @moduledoc false
 
-      def match?(number) do
+      def match?(unquote(number_prefix) <> _ = number) do
         Regex.match?(regex, number)
       end
+      def match?(_), do: false
 
       def builder(number) do
         [[_, code, area, number]] = Regex.scan(regex,number)
@@ -81,9 +83,9 @@ defmodule Helper.Country do
     end
   end
 
-  defp build_matcher(matcher) do
+  defp build_matcher(matcher, number_prefix) do
     case matcher do
-      :regex -> regex_matcher
+      :regex -> regex_matcher(number_prefix)
       :modules -> modules_matcher
       true ->
         raise ArgumentError, "You can only match against :regex or :modules, passed #{inspect matcher}"