--- a/lib/phone/ad.ex
+++ b/lib/phone/ad.ex
@@ -1,8 +1,7 @@
 defmodule Phone.AD do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(376)()(.{6})/
   field :country, "Andorra"
   field :a2, "AD"
   field :a3, "AND"
-  match :regex
 end