--- a/lib/phone/ad.ex
+++ b/lib/phone/ad.ex
@@ -1,5 +1,6 @@
 defmodule Phone.AD do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "376"
   field :regex, ~r/^(376)()(.{6})/
   field :country, "Andorra"
   field :a2, "AD"