--- a/lib/phone/ae.ex
+++ b/lib/phone/ae.ex
@@ -1,8 +1,7 @@
 defmodule Phone.AE do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(971)(.)(.{7})/
   field :country, "United Arab Emirates"
   field :a2, "AE"
   field :a3, "ARE"
-  match :regex
 end