--- a/lib/phone/ae.ex
+++ b/lib/phone/ae.ex
@@ -1,5 +1,6 @@
 defmodule Phone.AE do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "971"
   field :regex, ~r/^(971)(.)(.{7})/
   field :country, "United Arab Emirates"
   field :a2, "AE"