--- a/lib/phone/af.ex
+++ b/lib/phone/af.ex
@@ -1,8 +1,7 @@
 defmodule Phone.AF do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(93)(..)(.{7})/
   field :country, "Afghanistan"
   field :a2, "AF"
   field :a3, "AFG"
-  match :regex
 end