--- a/lib/phone/af.ex
+++ b/lib/phone/af.ex
@@ -1,5 +1,6 @@
 defmodule Phone.AF do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "93"
   field :regex, ~r/^(93)(..)(.{7})/
   field :country, "Afghanistan"
   field :a2, "AF"