--- a/lib/phone/al.ex
+++ b/lib/phone/al.ex
@@ -1,8 +1,7 @@
 defmodule Phone.AL do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(355)()(.{7})/
   field :country, "Albania"
   field :a2, "AL"
   field :a3, "ALB"
-  match :regex
 end