--- a/lib/phone/al.ex
+++ b/lib/phone/al.ex
@@ -1,5 +1,6 @@
 defmodule Phone.AL do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "355"
   field :regex, ~r/^(355)()(.{7})/
   field :country, "Albania"
   field :a2, "AL"