--- a/lib/phone/am.ex
+++ b/lib/phone/am.ex
@@ -1,8 +1,7 @@
 defmodule Phone.AM do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(374)()(.{8})/
   field :country, "Armenia"
   field :a2, "AM"
   field :a3, "ARM"
-  match :regex
 end