--- a/lib/phone/am.ex
+++ b/lib/phone/am.ex
@@ -1,5 +1,6 @@
 defmodule Phone.AM do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "374"
   field :regex, ~r/^(374)()(.{8})/
   field :country, "Armenia"
   field :a2, "AM"