--- a/lib/phone/ao.ex
+++ b/lib/phone/ao.ex
@@ -1,8 +1,7 @@
 defmodule Phone.AO do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(244)()(9)/
   field :country, "Angola"
   field :a2, "AO"
   field :a3, "AGO"
-  match :regex
 end