--- a/lib/phone/ao.ex
+++ b/lib/phone/ao.ex
@@ -1,5 +1,6 @@
 defmodule Phone.AO do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "244"
   field :regex, ~r/^(244)()(9)/
   field :country, "Angola"
   field :a2, "AO"