--- a/lib/phone/ar.ex
+++ b/lib/phone/ar.ex
@@ -1,8 +1,7 @@
 defmodule Phone.AR do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(54)()(.{10})/
   field :country, "Argentina"
   field :a2, "AR"
   field :a3, "ARG"
-  match :regex
 end