--- a/lib/phone/ar.ex
+++ b/lib/phone/ar.ex
@@ -1,5 +1,6 @@
 defmodule Phone.AR do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "54"
   field :regex, ~r/^(54)()(.{10})/
   field :country, "Argentina"
   field :a2, "AR"