--- a/lib/phone/at.ex
+++ b/lib/phone/at.ex
@@ -1,8 +1,7 @@
 defmodule Phone.AT do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(43)()(.+)/
   field :country, "Austria"
   field :a2, "AT"
   field :a3, "AUT"
-  match :regex
 end