--- a/lib/phone/at.ex
+++ b/lib/phone/at.ex
@@ -1,5 +1,6 @@
 defmodule Phone.AT do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "43"
   field :regex, ~r/^(43)()(.+)/
   field :country, "Austria"
   field :a2, "AT"