--- a/lib/phone/au.ex
+++ b/lib/phone/au.ex
@@ -1,8 +1,7 @@
 defmodule Phone.AU do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(61)([1-478])(.{8})/
   field :country, "Australia"
   field :a2, "AU"
   field :a3, "AUS"
-  match :regex
 end