--- a/lib/phone/au.ex
+++ b/lib/phone/au.ex
@@ -1,5 +1,6 @@
 defmodule Phone.AU do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "61"
   field :regex, ~r/^(61)([1-478])(.{8})/
   field :country, "Australia"
   field :a2, "AU"