--- a/lib/phone/aw.ex
+++ b/lib/phone/aw.ex
@@ -1,8 +1,7 @@
 defmodule Phone.AW do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(297)()(.{7})/
   field :country, "Aruba"
   field :a2, "AW"
   field :a3, "ABW"
-  match :regex
 end