--- a/lib/phone/aw.ex
+++ b/lib/phone/aw.ex
@@ -1,5 +1,6 @@
 defmodule Phone.AW do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "297"
   field :regex, ~r/^(297)()(.{7})/
   field :country, "Aruba"
   field :a2, "AW"