--- a/lib/phone/az.ex
+++ b/lib/phone/az.ex
@@ -1,8 +1,7 @@
 defmodule Phone.AZ do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(994)()(.{9})/
   field :country, "Azerbeijan"
   field :a2, "AZ"
   field :a3, "AZE"
-  match :regex
 end