--- a/lib/phone/az.ex
+++ b/lib/phone/az.ex
@@ -1,5 +1,6 @@
 defmodule Phone.AZ do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "994"
   field :regex, ~r/^(994)()(.{9})/
   field :country, "Azerbeijan"
   field :a2, "AZ"