--- a/lib/phone/ba.ex
+++ b/lib/phone/ba.ex
@@ -1,8 +1,7 @@
 defmodule Phone.BA do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(387)(..)(.{5,6})/
   field :country, "Bosnia and Herzegovina"
   field :a2, "BA"
   field :a3, "BIH"
-  match :regex
 end