--- a/lib/phone/bd.ex
+++ b/lib/phone/bd.ex
@@ -1,8 +1,7 @@
 defmodule Phone.BD do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(880)()(.{10})/
   field :country, "Bangladesh"
   field :a2, "BD"
   field :a3, "BGD"
-  match :regex
 end