--- a/lib/phone/bd.ex
+++ b/lib/phone/bd.ex
@@ -1,5 +1,6 @@
 defmodule Phone.BD do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "880"
   field :regex, ~r/^(880)()(.{10})/
   field :country, "Bangladesh"
   field :a2, "BD"