--- a/lib/phone/be.ex
+++ b/lib/phone/be.ex
@@ -1,8 +1,7 @@
 defmodule Phone.BE do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(32)()(.{9})/
   field :country, "Belgium"
   field :a2, "BE"
   field :a3, "BEL"
-  match :regex
 end