--- a/lib/phone/be.ex
+++ b/lib/phone/be.ex
@@ -1,5 +1,6 @@
 defmodule Phone.BE do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "32"
   field :regex, ~r/^(32)()(.{9})/
   field :country, "Belgium"
   field :a2, "BE"