--- a/lib/phone/bg.ex
+++ b/lib/phone/bg.ex
@@ -1,8 +1,7 @@
 defmodule Phone.BG do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(359)()(.{8,9})/
   field :country, "Bulgaria"
   field :a2, "BG"
   field :a3, "BGR"
-  match :regex
 end