--- a/lib/phone/bg.ex
+++ b/lib/phone/bg.ex
@@ -1,5 +1,6 @@
 defmodule Phone.BG do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "359"
   field :regex, ~r/^(359)()(.{8,9})/
   field :country, "Bulgaria"
   field :a2, "BG"