--- a/lib/phone/bh.ex
+++ b/lib/phone/bh.ex
@@ -1,8 +1,7 @@
 defmodule Phone.BH do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(973)()(.{8})/
   field :country, "Bahrain"
   field :a2, "BH"
   field :a3, "BHR"
-  match :regex
 end