--- a/lib/phone/bh.ex
+++ b/lib/phone/bh.ex
@@ -1,5 +1,6 @@
 defmodule Phone.BH do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "973"
   field :regex, ~r/^(973)()(.{8})/
   field :country, "Bahrain"
   field :a2, "BH"