--- a/lib/phone/bi.ex
+++ b/lib/phone/bi.ex
@@ -1,8 +1,7 @@
 defmodule Phone.BI do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(257)()(.{8})/
   field :country, "Burundi"
   field :a2, "BI"
   field :a3, "BDI"
-  match :regex
 end