--- a/lib/phone/bi.ex
+++ b/lib/phone/bi.ex
@@ -1,5 +1,6 @@
 defmodule Phone.BI do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "257"
   field :regex, ~r/^(257)()(.{8})/
   field :country, "Burundi"
   field :a2, "BI"