--- a/lib/phone/bj.ex
+++ b/lib/phone/bj.ex
@@ -1,8 +1,7 @@
 defmodule Phone.BJ do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(229)()(.{8})/
   field :country, "Benin"
   field :a2, "BJ"
   field :a3, "BEN"
-  match :regex
 end