--- a/lib/phone/bj.ex
+++ b/lib/phone/bj.ex
@@ -1,5 +1,6 @@
 defmodule Phone.BJ do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "229"
   field :regex, ~r/^(229)()(.{8})/
   field :country, "Benin"
   field :a2, "BJ"