--- a/lib/phone/bn.ex
+++ b/lib/phone/bn.ex
@@ -1,8 +1,7 @@
 defmodule Phone.BN do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(673)()(.{7})/
   field :country, "Brunei"
   field :a2, "BN"
   field :a3, "BRN"
-  match :regex
 end