--- a/lib/phone/bn.ex
+++ b/lib/phone/bn.ex
@@ -1,5 +1,6 @@
 defmodule Phone.BN do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "673"
   field :regex, ~r/^(673)()(.{7})/
   field :country, "Brunei"
   field :a2, "BN"