--- a/lib/phone/bo.ex
+++ b/lib/phone/bo.ex
@@ -1,8 +1,7 @@
 defmodule Phone.BO do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(591)()(.{8})/
   field :country, "Bolivia"
   field :a2, "BO"
   field :a3, "BOL"
-  match :regex
 end