--- a/lib/phone/bo.ex
+++ b/lib/phone/bo.ex
@@ -1,5 +1,6 @@
 defmodule Phone.BO do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "591"
   field :regex, ~r/^(591)()(.{8})/
   field :country, "Bolivia"
   field :a2, "BO"