--- a/lib/phone/bt.ex
+++ b/lib/phone/bt.ex
@@ -1,8 +1,7 @@
 defmodule Phone.BT do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(975)()(.{7,8})/
   field :country, "Bhutan"
   field :a2, "BT"
   field :a3, "BTN"
-  match :regex
 end