--- a/lib/phone/bt.ex
+++ b/lib/phone/bt.ex
@@ -1,5 +1,6 @@
 defmodule Phone.BT do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "975"
   field :regex, ~r/^(975)()(.{7,8})/
   field :country, "Bhutan"
   field :a2, "BT"