--- a/lib/phone/bw.ex
+++ b/lib/phone/bw.ex
@@ -1,8 +1,7 @@
 defmodule Phone.BW do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(267)()(.{7})/
   field :country, "Botswana"
   field :a2, "BW"
   field :a3, "BWA"
-  match :regex
 end