--- a/lib/phone/bw.ex
+++ b/lib/phone/bw.ex
@@ -1,5 +1,6 @@
 defmodule Phone.BW do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "267"
   field :regex, ~r/^(267)()(.{7})/
   field :country, "Botswana"
   field :a2, "BW"