--- a/lib/phone/by.ex
+++ b/lib/phone/by.ex
@@ -1,8 +1,7 @@
 defmodule Phone.BY do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(375)()(.{9})/
   field :country, "Belarus"
   field :a2, "BY"
   field :a3, "BLR"
-  match :regex
 end