--- a/lib/phone/by.ex
+++ b/lib/phone/by.ex
@@ -1,5 +1,6 @@
 defmodule Phone.BY do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "375"
   field :regex, ~r/^(375)()(.{9})/
   field :country, "Belarus"
   field :a2, "BY"