--- a/lib/phone/bz.ex
+++ b/lib/phone/bz.ex
@@ -1,8 +1,7 @@
 defmodule Phone.BZ do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(501)()(.{7})/
   field :country, "Belize"
   field :a2, "BZ"
   field :a3, "BLZ"
-  match :regex
 end