--- a/lib/phone/bz.ex
+++ b/lib/phone/bz.ex
@@ -1,5 +1,6 @@
 defmodule Phone.BZ do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "501"
   field :regex, ~r/^(501)()(.{7})/
   field :country, "Belize"
   field :a2, "BZ"