--- a/lib/phone/cd.ex
+++ b/lib/phone/cd.ex
@@ -1,8 +1,7 @@
 defmodule Phone.CD do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(243)(.)(.{7})/
   field :country, "Democratic Republic of Congo"
   field :a2, "CD"
   field :a3, "COD"
-  match :regex
 end