--- a/lib/phone/cd.ex
+++ b/lib/phone/cd.ex
@@ -1,5 +1,6 @@
 defmodule Phone.CD do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "243"
   field :regex, ~r/^(243)(.)(.{7})/
   field :country, "Democratic Republic of Congo"
   field :a2, "CD"