--- a/lib/phone/cf.ex
+++ b/lib/phone/cf.ex
@@ -1,8 +1,7 @@
 defmodule Phone.CF do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(236)()(.{8})/
   field :country, "Central African Republic"
   field :a2, "CF"
   field :a3, "CAF"
-  match :regex
 end