--- a/lib/phone/cf.ex
+++ b/lib/phone/cf.ex
@@ -1,5 +1,6 @@
 defmodule Phone.CF do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "236"
   field :regex, ~r/^(236)()(.{8})/
   field :country, "Central African Republic"
   field :a2, "CF"