--- a/lib/phone/cg.ex
+++ b/lib/phone/cg.ex
@@ -1,8 +1,7 @@
 defmodule Phone.CG do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(242)(.{4})(.{5})/
   field :country, "Congo"
   field :a2, "CG"
   field :a3, "COG"
-  match :regex
 end