--- a/lib/phone/cg.ex
+++ b/lib/phone/cg.ex
@@ -1,5 +1,6 @@
 defmodule Phone.CG do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "242"
   field :regex, ~r/^(242)(.{4})(.{5})/
   field :country, "Congo"
   field :a2, "CG"