--- a/lib/phone/ch.ex
+++ b/lib/phone/ch.ex
@@ -1,8 +1,7 @@
 defmodule Phone.CH do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(41)()(.{9,10})/
   field :country, "Switzerland"
   field :a2, "CH"
   field :a3, "CHE"
-  match :regex
 end