--- a/lib/phone/ch.ex
+++ b/lib/phone/ch.ex
@@ -1,5 +1,6 @@
 defmodule Phone.CH do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "41"
   field :regex, ~r/^(41)()(.{9,10})/
   field :country, "Switzerland"
   field :a2, "CH"