--- a/lib/phone/ci.ex
+++ b/lib/phone/ci.ex
@@ -1,8 +1,7 @@
 defmodule Phone.CI do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(225)(..)(.{6})/
   field :country, "Ivory Coast"
   field :a2, "CI"
   field :a3, "CIV"
-  match :regex
 end