--- a/lib/phone/ci.ex
+++ b/lib/phone/ci.ex
@@ -1,5 +1,6 @@
 defmodule Phone.CI do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "225"
   field :regex, ~r/^(225)(..)(.{6})/
   field :country, "Ivory Coast"
   field :a2, "CI"