--- a/lib/phone/ck.ex
+++ b/lib/phone/ck.ex
@@ -1,8 +1,7 @@
 defmodule Phone.CK do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(682)()(.{5})/
   field :country, "Cook Islands"
   field :a2, "CK"
   field :a3, "COK"
-  match :regex
 end