--- a/lib/phone/ck.ex
+++ b/lib/phone/ck.ex
@@ -1,5 +1,6 @@
 defmodule Phone.CK do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "682"
   field :regex, ~r/^(682)()(.{5})/
   field :country, "Cook Islands"
   field :a2, "CK"