--- a/lib/phone/cl.ex
+++ b/lib/phone/cl.ex
@@ -1,8 +1,7 @@
 defmodule Phone.CL do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(56)()(.{9})/
   field :country, "Chile"
   field :a2, "CL"
   field :a3, "CHL"
-  match :regex
 end