--- a/lib/phone/cl.ex
+++ b/lib/phone/cl.ex
@@ -1,5 +1,6 @@
 defmodule Phone.CL do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "56"
   field :regex, ~r/^(56)()(.{9})/
   field :country, "Chile"
   field :a2, "CL"