--- a/lib/phone/cm.ex
+++ b/lib/phone/cm.ex
@@ -1,8 +1,7 @@
 defmodule Phone.CM do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(237)()(.{8})/
   field :country, "Cameroon"
   field :a2, "CM"
   field :a3, "CMR"
-  match :regex
 end