--- a/lib/phone/cm.ex
+++ b/lib/phone/cm.ex
@@ -1,5 +1,6 @@
 defmodule Phone.CM do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "237"
   field :regex, ~r/^(237)()(.{8})/
   field :country, "Cameroon"
   field :a2, "CM"