--- a/lib/phone/cn.ex
+++ b/lib/phone/cn.ex
@@ -1,8 +1,7 @@
 defmodule Phone.CN do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(86)()(.+)/
   field :country, "China"
   field :a2, "CN"
   field :a3, "CHN"
-  match :regex
 end