--- a/lib/phone/cn.ex
+++ b/lib/phone/cn.ex
@@ -1,5 +1,6 @@
 defmodule Phone.CN do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "86"
   field :regex, ~r/^(86)()(.+)/
   field :country, "China"
   field :a2, "CN"