--- a/lib/phone/co.ex
+++ b/lib/phone/co.ex
@@ -1,8 +1,7 @@
 defmodule Phone.CO do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(57)()(.+)/
   field :country, "Colombia"
   field :a2, "CO"
   field :a3, "COL"
-  match :regex
 end