--- a/lib/phone/cr.ex
+++ b/lib/phone/cr.ex
@@ -1,8 +1,7 @@
 defmodule Phone.CR do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(506)()(.{8})/
   field :country, "Costa Rica"
   field :a2, "CR"
   field :a3, "CRI"
-  match :regex
 end