--- a/lib/phone/cr.ex
+++ b/lib/phone/cr.ex
@@ -1,5 +1,6 @@
 defmodule Phone.CR do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "506"
   field :regex, ~r/^(506)()(.{8})/
   field :country, "Costa Rica"
   field :a2, "CR"