--- a/lib/phone/cu.ex
+++ b/lib/phone/cu.ex
@@ -1,8 +1,7 @@
 defmodule Phone.CU do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(53)()(.{8})/
   field :country, "Cuba"
   field :a2, "CU"
   field :a3, "CUB"
-  match :regex
 end