--- a/lib/phone/cu.ex
+++ b/lib/phone/cu.ex
@@ -1,5 +1,6 @@
 defmodule Phone.CU do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "53"
   field :regex, ~r/^(53)()(.{8})/
   field :country, "Cuba"
   field :a2, "CU"