--- a/lib/phone/cv.ex
+++ b/lib/phone/cv.ex
@@ -1,8 +1,7 @@
 defmodule Phone.CV do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(238)()(.{7})/
   field :country, "Cape Verde"
   field :a2, "CV"
   field :a3, "CPV"
-  match :regex
 end