--- a/lib/phone/cv.ex
+++ b/lib/phone/cv.ex
@@ -1,5 +1,6 @@
 defmodule Phone.CV do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "238"
   field :regex, ~r/^(238)()(.{7})/
   field :country, "Cape Verde"
   field :a2, "CV"