--- a/lib/phone/cw.ex
+++ b/lib/phone/cw.ex
@@ -1,8 +1,7 @@
 defmodule Phone.CW do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(599)(9)(.{7})/
   field :country, "Curaçao"
   field :a2, "CW"
   field :a3, "CUW"
-  match :regex
 end