--- a/lib/phone/cw.ex
+++ b/lib/phone/cw.ex
@@ -1,5 +1,6 @@
 defmodule Phone.CW do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "599"
   field :regex, ~r/^(599)(9)(.{7})/
   field :country, "Curaçao"
   field :a2, "CW"