--- a/lib/phone/cy.ex
+++ b/lib/phone/cy.ex
@@ -1,8 +1,7 @@
 defmodule Phone.CY do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(357)()(.{7,8})/
   field :country, "Cyprus"
   field :a2, "CY"
   field :a3, "CYP"
-  match :regex
 end