--- a/lib/phone/cy.ex
+++ b/lib/phone/cy.ex
@@ -1,5 +1,6 @@
 defmodule Phone.CY do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "357"
   field :regex, ~r/^(357)()(.{7,8})/
   field :country, "Cyprus"
   field :a2, "CY"