--- a/lib/phone/cz.ex
+++ b/lib/phone/cz.ex
@@ -1,8 +1,7 @@
 defmodule Phone.CZ do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(420)()(.{9})/
   field :country, "Czech Republic"
   field :a2, "CZ"
   field :a3, "CZE"
-  match :regex
 end