--- a/lib/phone/cz.ex
+++ b/lib/phone/cz.ex
@@ -1,5 +1,6 @@
 defmodule Phone.CZ do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "420"
   field :regex, ~r/^(420)()(.{9})/
   field :country, "Czech Republic"
   field :a2, "CZ"