--- a/lib/phone/de.ex
+++ b/lib/phone/de.ex
@@ -1,8 +1,7 @@
 defmodule Phone.DE do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(49)()(.{10,11})/
   field :country, "Germany"
   field :a2, "DE"
   field :a3, "DEU"
-  match :regex
 end