--- a/lib/phone/de.ex
+++ b/lib/phone/de.ex
@@ -1,5 +1,6 @@
 defmodule Phone.DE do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "49"
   field :regex, ~r/^(49)()(.{10,11})/
   field :country, "Germany"
   field :a2, "DE"