--- a/lib/phone/dj.ex
+++ b/lib/phone/dj.ex
@@ -1,8 +1,7 @@
 defmodule Phone.DJ do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(253)()(.{8})/
   field :country, "Djibouti"
   field :a2, "DJ"
   field :a3, "DJI"
-  match :regex
 end