--- a/lib/phone/dj.ex
+++ b/lib/phone/dj.ex
@@ -1,5 +1,6 @@
 defmodule Phone.DJ do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "253"
   field :regex, ~r/^(253)()(.{8})/
   field :country, "Djibouti"
   field :a2, "DJ"