--- a/lib/phone/dk.ex
+++ b/lib/phone/dk.ex
@@ -1,8 +1,7 @@
 defmodule Phone.DK do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(45)()(.{8})/
   field :country, "Denmark"
   field :a2, "DK"
   field :a3, "DNK"
-  match :regex
 end