--- a/lib/phone/dk.ex
+++ b/lib/phone/dk.ex
@@ -1,5 +1,6 @@
 defmodule Phone.DK do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "45"
   field :regex, ~r/^(45)()(.{8})/
   field :country, "Denmark"
   field :a2, "DK"