--- a/lib/phone/dz.ex
+++ b/lib/phone/dz.ex
@@ -1,8 +1,7 @@
 defmodule Phone.DZ do
-   use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(213)()(.{8})/
   field :country, "Algeria"
   field :a2, "DZ"
   field :a3, "DZA"
-  match :regex
 end