--- a/lib/phone/dz.ex
+++ b/lib/phone/dz.ex
@@ -1,5 +1,6 @@
 defmodule Phone.DZ do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "213"
   field :regex, ~r/^(213)()(.{8})/
   field :country, "Algeria"
   field :a2, "DZ"