--- a/lib/phone/ec.ex
+++ b/lib/phone/ec.ex
@@ -1,8 +1,7 @@
 defmodule Phone.EC do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(593)(..)(.{7})/
   field :country, "Ecuador"
   field :a2, "EC"
   field :a3, "ECU"
-  match :regex
 end