--- a/lib/phone/ec.ex
+++ b/lib/phone/ec.ex
@@ -1,5 +1,6 @@
 defmodule Phone.EC do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "593"
   field :regex, ~r/^(593)(..)(.{7})/
   field :country, "Ecuador"
   field :a2, "EC"