--- a/lib/phone/ee.ex
+++ b/lib/phone/ee.ex
@@ -1,8 +1,7 @@
 defmodule Phone.EE do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(372)()(.{7,8})/
   field :country, "Estonia"
   field :a2, "EE"
   field :a3, "EST"
-  match :regex
 end