--- a/lib/phone/ee.ex
+++ b/lib/phone/ee.ex
@@ -1,5 +1,6 @@
 defmodule Phone.EE do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "372"
   field :regex, ~r/^(372)()(.{7,8})/
   field :country, "Estonia"
   field :a2, "EE"