--- a/lib/phone/eg.ex
+++ b/lib/phone/eg.ex
@@ -1,8 +1,7 @@
 defmodule Phone.EG do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(20)()(.+)/
   field :country, "Egypt"
   field :a2, "EG"
   field :a3, "EGY"
-  match :regex
 end