--- a/lib/phone/eg.ex
+++ b/lib/phone/eg.ex
@@ -1,5 +1,6 @@
 defmodule Phone.EG do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "20"
   field :regex, ~r/^(20)()(.+)/
   field :country, "Egypt"
   field :a2, "EG"