--- a/lib/phone/er.ex
+++ b/lib/phone/er.ex
@@ -1,8 +1,7 @@
 defmodule Phone.ER do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(291)(.)(.{6})/
   field :country, "Eritrea"
   field :a2, "ER"
   field :a3, "ERI"
-  match :regex
 end