--- a/lib/phone/er.ex
+++ b/lib/phone/er.ex
@@ -1,5 +1,6 @@
 defmodule Phone.ER do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "291"
   field :regex, ~r/^(291)(.)(.{6})/
   field :country, "Eritrea"
   field :a2, "ER"