--- a/lib/phone/es.ex
+++ b/lib/phone/es.ex
@@ -1,8 +1,7 @@
 defmodule Phone.ES do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(34)()(.{9})/
   field :country, "Spain"
   field :a2, "ES"
   field :a3, "ESP"
-  match :regex
 end