--- a/lib/phone/es.ex
+++ b/lib/phone/es.ex
@@ -1,5 +1,6 @@
 defmodule Phone.ES do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "34"
   field :regex, ~r/^(34)()(.{9})/
   field :country, "Spain"
   field :a2, "ES"