--- a/lib/phone/et.ex
+++ b/lib/phone/et.ex
@@ -1,8 +1,7 @@
 defmodule Phone.ET do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(251)(..)(.{7})/
   field :country, "Ethiopia"
   field :a2, "ET"
   field :a3, "ETH"
-  match :regex
 end