--- a/lib/phone/et.ex
+++ b/lib/phone/et.ex
@@ -1,5 +1,6 @@
 defmodule Phone.ET do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "251"
   field :regex, ~r/^(251)(..)(.{7})/
   field :country, "Ethiopia"
   field :a2, "ET"