--- a/lib/phone/fi.ex
+++ b/lib/phone/fi.ex
@@ -1,8 +1,7 @@
 defmodule Phone.FI do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(358)()(.+)/
   field :country, "Finland"
   field :a2, "FI"
   field :a3, "FIN"
-  match :regex
 end