--- a/lib/phone/fi.ex
+++ b/lib/phone/fi.ex
@@ -1,5 +1,6 @@
 defmodule Phone.FI do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "358"
   field :regex, ~r/^(358)()(.+)/
   field :country, "Finland"
   field :a2, "FI"