--- a/lib/phone/fj.ex
+++ b/lib/phone/fj.ex
@@ -1,8 +1,7 @@
 defmodule Phone.FJ do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(679)()(.{7})/
   field :country, "Fiji"
   field :a2, "FJ"
   field :a3, "FJI"
-  match :regex
 end