--- a/lib/phone/fj.ex
+++ b/lib/phone/fj.ex
@@ -1,5 +1,6 @@
 defmodule Phone.FJ do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "679"
   field :regex, ~r/^(679)()(.{7})/
   field :country, "Fiji"
   field :a2, "FJ"