--- a/lib/phone/fm.ex
+++ b/lib/phone/fm.ex
@@ -1,8 +1,7 @@
 defmodule Phone.FM do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(691)()(.{7})/
   field :country, "Micronesia"
   field :a2, "FM"
   field :a3, "FSM"
-  match :regex
 end