--- a/lib/phone/fm.ex
+++ b/lib/phone/fm.ex
@@ -1,5 +1,6 @@
 defmodule Phone.FM do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "691"
   field :regex, ~r/^(691)()(.{7})/
   field :country, "Micronesia"
   field :a2, "FM"