--- a/lib/phone/fo.ex
+++ b/lib/phone/fo.ex
@@ -1,8 +1,7 @@
 defmodule Phone.FO do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(298)()(.{6})/
   field :country, "Faroe Islands"
   field :a2, "FO"
   field :a3, "FRO"
-  match :regex
 end