--- a/lib/phone/fo.ex
+++ b/lib/phone/fo.ex
@@ -1,5 +1,6 @@
 defmodule Phone.FO do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "298"
   field :regex, ~r/^(298)()(.{6})/
   field :country, "Faroe Islands"
   field :a2, "FO"