--- a/lib/phone/fr.ex
+++ b/lib/phone/fr.ex
@@ -1,8 +1,7 @@
 defmodule Phone.FR do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(33)()(.{9})/
   field :country, "France"
   field :a2, "FR"
   field :a3, "FRA"
-  match :regex
 end