--- a/lib/phone/fr.ex
+++ b/lib/phone/fr.ex
@@ -1,5 +1,6 @@
 defmodule Phone.FR do
-  use Helper.Country, match: :regex
+  use Helper.Country, match: :regex,
+    number_prefix: "33"
   field :regex, ~r/^(33)()(.{9})/
   field :country, "France"
   field :a2, "FR"