--- a/lib/phone/ga.ex
+++ b/lib/phone/ga.ex
@@ -1,8 +1,7 @@
 defmodule Phone.GA do
-  use Helper.Country
+  use Helper.Country, match: :regex
   field :regex, ~r/^(241)()(.{7})/
   field :country, "Gabon"
   field :a2, "GA"
   field :a3, "GAB"
-  match :regex
 end