… CONTRIBUTING.md
… bench/phone_bench.exs
… lib/phone/gb.ex
… lib/phone/ge.ex
… lib/phone/gf.ex
… lib/phone/gh.ex
… lib/phone/gi.ex
… lib/phone/gl.ex
… lib/phone/gm.ex
… lib/phone/gn.ex
… lib/phone/gq.ex
… lib/phone/gr.ex
… lib/phone/gt.ex
… lib/phone/gw.ex
… lib/phone/gy.ex
… lib/phone/hk.ex
… lib/phone/hn.ex
… lib/phone/hr.ex
… lib/phone/ht.ex
… lib/phone/hu.ex
… lib/phone/id.ex
… lib/phone/ie.ex
… lib/phone/il.ex
… lib/phone/in.ex
… lib/phone/io.ex
… lib/phone/iq.ex
… lib/phone/ir.ex
… lib/phone/is.ex
… lib/phone/it.ex
… lib/phone/jo.ex