I think we can achieve a lot of performance gains in this library by using pattern matching to quickly narrow down the list of possible phone number parsers and then using regular expressions to extract the necessary data and parse the phone number. I tried the following benchmark earlier today to compare pattern matching vs regular expressions and found there's a huge speed difference.

```elixir
defmodule PhoneBench do
  use Benchfella

  @andorra_phone_number "376123456"

  def phone_number_matches?("376"<>_), do: true
  def phone_number_matches?(_), do: false

  bench "pattern match with prefix" do
    true = phone_number_matches?(@andorra_phone_number)
  end

  bench "match with regex" do
    true = Regex.match?(~r/^(376)()(.{6})/, @andorra_phone_number)
  end
end
```

```
## PhoneBench
benchmark name                              iterations   average time
pattern match with prefix                    100000000   0.04 µs/op
match with regex                               1000000   1.24 µs/op
```

That makes me think if we can leverage elixir's pattern matching, we have a tone of room for performance improvements.

This is my first try at using pattern matching, and I think there are a lot more gains we can make. I changed how you set the the match type for each country module (to eventually allow for overriding functions) and added a `number_prefix` option that helps with the pattern matching in the generated `match?` function. I would recommend looking at the commits one at a time and letting me know what you think.
Here's the before and after performance metrics.

```
Before

## PhoneBench
benchmark name                              iterations   average time
Phone.parse/1 with an Andorra phone number        5000   673.05 µs/op
Phone.parse/1 with a Zimbabwe phone number        5000   691.99 µs/op
Phone.parse/1 with an NANP phone number           1000   1047.04 µs/op

With these changes

## PhoneBench
benchmark name                              iterations   average time
Phone.parse/1 with an Andorra phone number       10000   294.79 µs/op
Phone.parse/1 with a Zimbabwe phone number       10000   298.49 µs/op
Phone.parse/1 with an NANP phone number           5000   713.80 µs/op

Difference

## PhoneBench
Phone.parse/1 with a Zimbabwe phone number    0.43
Phone.parse/1 with an Andorra phone number    0.44
Phone.parse/1 with an NANP phone number       0.68
```

I'm curious to see if we could have even bigger gains if we could move the phone number parsing into one giant (generated) parse function with one definition for each prefix. I'm not sure how we'd do it, but it would be interesting to see how fast we can make this library.