@aaronrenner 2016-11-19
Regarding doing phone number parsing with multiple function definitions, the MIME library shows a great example of how to do this: https://github.com/elixir-lang/mime/blob/master/lib/mime.ex#L93-L100. I think I'll look into that and see if I can put together something.