@fcevado 2016-11-19
Great to know that pattern matching even with string partials are faster than regex. I'll take a look at it and try to come up with some strategy to have less pattern matching possible.