@fcevado 2016-11-19
I've done some tests with pure pattern matching and pure regex match, I've come up with this results:
```
benchmark name    iterations   average time 
pattern matching   100000000   0.04 µs/op
pure regex          10000000   0.72 µs/op
```
I did it with only one case of pattern matching, I think if that number increase the performance would drop, but the difference is huge. I thought of changing the `Phone.Countries` module, so we could pattern match to identify the country instead of the current iteration way. Problems would come with some countries that have the same international code, for example Russia and Kazakhstan. But I think that can be solved with the iteration method only for those cases.
I'll take your commit that add benchmarks and try to come up with something.