@aaronrenner 2016-11-19
Ideally I'm thinking we could use metaprogramming to generate one big parse function with multiple definitions like this.

```elixir
def parse("1970" <> <<number::byte_size(7)>>) do
  # Run regex just to be sure
  Regex.match?(...)
  {:ok, %{a2: "US", a3: "USA", country: "United States", international_code: "1", area_code: "970", number: number, area_abbreviation: "CO", area_type: "state", area_name: "Colorado"}}
end
def parse("1303" <> <<number::byte_size(7)>>) do
  # Run regex just to be sure
  Regex.match?(...)
  {:ok, %{a2: "US", a3: "USA", country: "United States", international_code: "1", area_code: "303", number: number, area_abbreviation: "CO", area_type: "state", area_name: "Colorado"}}
end
```
My thought is calling one big function will be faster than calling a function for each country (but we should benchmark that to see if it's true). The trick would be storing all of this data so you could generate this function at compile time. Also, I wonder if it would be faster to generate separate function definitions for each area code in a state or use the in guard clause like this:

```elixir
def parse("1"<<area_code::byte_size(3)>> <> <<number::byte_size(7)>>) when area_code in ~w(303 719 720 970) do
  # Run regex just to be sure
  Regex.match?(...)
  {:ok, %{a2: "US", a3: "USA", country: "United States", international_code: "1", area_code: area_code, number: number, area_abbreviation: "CO", area_type: "state", area_name: "Colorado"}}
end
```
This sure is exciting to find room for such a huge performance gain. I don't think we have to do all of the checking via pattern matching, just enough to quickly exclude the other countries so we only have to run 1 regex.

Thanks again for looking at this. Please let me know how I can help.