@fcevado 2016-11-19
I understand that performance is a huge issue, but I think the main objective is try to keep the lib easy to contribute(phone rules use to change a lot, even more in small countries). My first idea was to change the `Phone.Countries` module, so it would have a `build/1` and `build/2` function to pattern match every country code and call their modules directly. I'm thinking of doing it too with the `Phone.NANP` area codes too(but just for identify the countries) since it has a more stable numbering plan, if we can keep it at hundreds of µs is a good thing.