@fcevado 2016-11-19
@aaronrenner if you agree and want to spend more time with that, we could do it both ways, you do it the way you said and I create a pr with the way I'm thinking, and we compare the performances and readability of the code.
I think it would help both of us to learn how to achieve better performances understanding how it works on the BEAM.