@aaronrenner 2016-11-21
@fcevado I'd be happy to spend a little more time on this so we can try it both ways. Looking forward to the learning experience! 😄