@fcevado 2017-04-02
@aaronrenner I'll close those PRs, I worked on that this weekend and I believe that i reached it, follows my results:
```
## PhoneBench
benchmark name                              iterations   average time 
Phone.parse/1 with an Andorra phone number      500000   7.93 µs/op
Phone.parse/1 with a Zimbabwe phone number      500000   7.99 µs/op
Phone.parse/1 with an NANP phone number         200000   9.95 µs/op
```
In a counterpart the complexity was thrown to the compiler, the compilation time is really big.